+++ app/src/main/java/org/simple/clinic/home/overdue/OverdueModel.kt
@@ -1,0 +1,12 @@
+package org.simple.clinic.home.overdue
+
+import android.os.Parcelable
+import kotlinx.android.parcel.Parcelize
+
+@Parcelize
+class OverdueModel: Parcelable {
+
+  companion object {
+    fun create(): OverdueModel = OverdueModel()
+  }
+}