+++ app/src/main/java/org/simple/clinic/home/overdue/OverdueScreen.kt
@@ -1,18 +1,23 @@
 package org.simple.clinic.home.overdue
 
 import android.content.Context
+import android.os.Parcelable
 import android.util.AttributeSet
 import android.widget.RelativeLayout
 import androidx.appcompat.app.AppCompatActivity
 import androidx.recyclerview.widget.LinearLayoutManager
-import com.jakewharton.rxbinding2.view.RxView
+import com.jakewharton.rxbinding3.view.detaches
 import io.reactivex.Observable
+import io.reactivex.rxkotlin.ofType
 import kotlinx.android.synthetic.main.screen_overdue.view.*
+import org.simple.clinic.ReportAnalyticsEvents
 import org.simple.clinic.bindUiToController
 import org.simple.clinic.contactpatient.ContactPatientBottomSheet
-import org.simple.clinic.main.TheActivity
+import org.simple.clinic.di.injector
+import org.simple.clinic.mobius.MobiusDelegate
 import org.simple.clinic.router.screen.ScreenRouter
 import org.simple.clinic.util.UserClock
+import org.simple.clinic.util.unsafeLazy
 import org.simple.clinic.widgets.ItemAdapter
 import org.simple.clinic.widgets.ScreenDestroyed
 import org.simple.clinic.widgets.visibleOrGone
@@ -21,7 +26,10 @@ import java.util.UUID
 import javax.inject.Inject
 import javax.inject.Named
 
-class OverdueScreen(context: Context, attrs: AttributeSet) : RelativeLayout(context, attrs) {
+class OverdueScreen(
+    context: Context,
+    attrs: AttributeSet
+) : RelativeLayout(context, attrs), OverdueUi {
 
   @Inject
   lateinit var activity: AppCompatActivity
@@ -38,7 +46,33 @@ class OverdueScreen(context: Context, attrs: AttributeSet) : RelativeLayout(cont
   @field:[Inject Named("full_date")]
   lateinit var dateFormatter: DateTimeFormatter
 
+  @Inject
+  lateinit var effectHandlerFactory: OverdueEffectHandler.Factory
+
   private val overdueListAdapter = ItemAdapter(OverdueAppointmentRow.DiffCallback())
+
+  private val events by unsafeLazy {
+    Observable
+        .merge(
+            screenCreates(),
+            overdueListAdapter.itemEvents
+        )
+        .compose(ReportAnalyticsEvents())
+        .share()
+  }
+
+  private val delegate by unsafeLazy {
+    val uiRenderer = OverdueUiRenderer(this)
+
+    MobiusDelegate.forView(
+        events = events.ofType(),
+        defaultModel = OverdueModel.create(),
+        update = OverdueUpdate(),
+        effectHandler = effectHandlerFactory.create(this).build(),
+        init = OverdueInit(),
+        modelUpdateListener = uiRenderer::render
+    )
+  }
 
   override fun onFinishInflate() {
     super.onFinishInflate()
@@ -46,29 +80,42 @@ class OverdueScreen(context: Context, attrs: AttributeSet) : RelativeLayout(cont
       return
     }
 
-    TheActivity.component.inject(this)
+    context.injector<Injector>().inject(this)
 
     overdueRecyclerView.adapter = overdueListAdapter
     overdueRecyclerView.layoutManager = LinearLayoutManager(context)
 
-    val screenDestroys = RxView
-        .detaches(this)
-        .map { ScreenDestroyed() }
+    val screenDestroys = detaches().map { ScreenDestroyed() }
 
     bindUiToController(
         ui = this,
-        events = Observable.merge(
-            screenCreates(),
-            overdueListAdapter.itemEvents
-        ),
+        events = events,
         controller = controller,
         screenDestroys = screenDestroys
     )
   }
 
+  override fun onAttachedToWindow() {
+    super.onAttachedToWindow()
+    delegate.start()
+  }
+
+  override fun onDetachedFromWindow() {
+    delegate.stop()
+    super.onDetachedFromWindow()
+  }
+
+  override fun onSaveInstanceState(): Parcelable? {
+    return delegate.onSaveInstanceState(super.onSaveInstanceState())
+  }
+
+  override fun onRestoreInstanceState(state: Parcelable?) {
+    super.onRestoreInstanceState(delegate.onRestoreInstanceState(state))
+  }
+
   private fun screenCreates() = Observable.just(OverdueScreenCreated())
 
-  fun updateList(overdueAppointments: List<OverdueAppointment>, isDiabetesManagementEnabled: Boolean) {
+  override fun updateList(overdueAppointments: List<OverdueAppointment>, isDiabetesManagementEnabled: Boolean) {
     overdueListAdapter.submitList(OverdueAppointmentRow.from(
         appointments = overdueAppointments,
         clock = userClock,
@@ -77,12 +124,16 @@ class OverdueScreen(context: Context, attrs: AttributeSet) : RelativeLayout(cont
     ))
   }
 
-  fun handleEmptyList(isEmpty: Boolean) {
+  override fun handleEmptyList(isEmpty: Boolean) {
     viewForEmptyList.visibleOrGone(isEmpty)
     overdueRecyclerView.visibleOrGone(isEmpty.not())
   }
 
-  fun openPhoneMaskBottomSheet(patientUuid: UUID) {
+  override fun openPhoneMaskBottomSheet(patientUuid: UUID) {
     activity.startActivity(ContactPatientBottomSheet.intent(context, patientUuid))
+  }
+
+  interface Injector {
+    fun inject(target: OverdueScreen)
   }
 }