+++ CHANGELOG.md
@@ -4,8 +4,8 @@
 ###Internal
 - Migrated `LoginPinScreen` to Mobius
 - [In Progress: 24th Jul 2020] Extract facility selection UI into a separate component
-- [In Progress: 24th Jul 2020] Migrate `OverdueScreen` to Mobius
 - [In Progress: 28th Jul 2020] Migrate `ForgotPinCreateNewPinScreen` to Mobius
+- [In Progress: 28th Jul 2020] Migrate `OverdueScreen` to Mobius
 
 ## On Demo
 ### Features